requirements.txt
pypinyin
fabric3
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cssselect==1.0.3
cycler==0.10.0
Django==2.0.1
django-cors-headers==2.1.0
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itypes==1.1.0
jieba==0.39
Jinja2==2.10
lxml==4.1.1
Markdown==2.6.10
MarkupSafe==1.0
matplotlib==2.1.1
mysqlclient==1.3.12
numpy==1.13.3
olefile==0.44
openapi-codec==1.3.2
opencv-python==3.4.0.12
pandas==0.22.0
parsel==1.3.1
Pillow==5.0.0
pycparser==2.18
PyDispatcher==2.0.5
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
selenium==3.8.0
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
w3lib==1.19.0
zope.interface==4.4.3
